+++ composer.json
@@ -22,12 +22,8 @@
     "license": "MIT",
     "authors": [
         {
-            "name": "Olanipekun Femi",
-            "email": "iolufemi@ymail.com"
-        },
-        {
-            "name": "Emereuwaonu Eze",
-            "email": "emereuwaonueze@gmail.com"
+            "name": "Olaobaju Abraham",
+            "email": "olaobajua@gmail.com"
         }
     ]
 }
