--- lib/Qr.php
@@ -1,1 +1,0 @@
-
