+++ processPayment.php
@@ -5,8 +5,8 @@ session_start() ;
 // Prevent direct access to this class
 define("BASEPATH", 1);
 
-include('lib/rave.php');
-include('lib/raveEventHandlerInterface.php');
+include('library/rave.php');
+include('library/raveEventHandlerInterface.php');
 
 use Flutterwave\Rave;
 use Flutterwave\EventHandlerInterface;
